Post-extraction filtration based on URL structure
Implemented a post-extraction filtration system based on a regex whitelist. URLs without backslashes at the end and at least three words separated by hyphens are considered valid article URLs. The necessary tests have also been implemented, as well as an export function that saves the dataframes as CSVs with URL and datetime in the name.