Successful bypass of pur abo barrier
All tests are running successfully. Consent button to proceed without the pur abo is found (via iframe), clicked and the driver waits until the homepage is loaded before closing. This sets the stage for connection of this function with the extraction pipelines.
Compatibility with the GeckoDriver currently nonexistent, might be nice to have in the future.