Successful Click Implementation
Opened up a new branch due to issues with the previous one. The following things have been done: print() has been traded in for log.info(); all functions can now be run in the command line. Tests have not yet been adjusted and click functions just call the actual functions - this will not be the case in the future and is currently just to not open up too many roadblocks at once in the process.