Added formatting rules for comments and texts
To avoid breaking CSVs when extracting text and comments, the text in the values 'comments' and 'text' has been adjusted. I've left paragraph marks in for later navigation of the text - they can be be deleted easily, though, so no worries there. Also: one run of get-data is now executed as soon as the timer is set, so that the user doesn't have to wait X hours to see results.