Stabilized data collection
Various quality of scraping updates: added a popup clicker (which emulates the functionality of the now-removed pur abo pipeline) and adjusted some selenium-based data collection processes. Runs smoother and faster now, as well as more reliable (as proven by custom homepages, which will be added to default in due time).